requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
pydantic>=2.10.0
python-dotenv>=1.0.1
supabase>=2.8.1
anthropic>=0.39.0
duckdb>=1.1.3
pandas>=2.2.3
jinja2>=3.1.4
python-multipart>=0.0.12

